refactor: Theme 컨트롤러 관리자 기능 분리
