+++ src/main/java/roomescape/domain/theme/admin/AdminThemeController.java
@@ -1,0 +1,57 @@
+package roomescape.domain.theme.admin;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.Valid;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.theme.ThemeService;
+import roomescape.domain.theme.admin.dto.AdminThemeResponse;
+import roomescape.domain.theme.admin.dto.CreateThemeRequest;
+import roomescape.domain.theme.admin.dto.CreateThemeResponse;
+import roomescape.support.auth.AdminRequestValidator;
+
+@RestController
+@RequiredArgsConstructor
+public class AdminThemeController {
+
+    private final ThemeService themeService;
+    private final AdminRequestValidator validator;
+
+    @GetMapping("/admin/themes")
+    public ResponseEntity<List<AdminThemeResponse>> getAllThemeForAdmin(HttpServletRequest request) {
+        if (validator.isUnauthorized(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        List<AdminThemeResponse> response = themeService.getAllThemeForAdmin();
+        return ResponseEntity.ok(response);
+    }
+
+    @PostMapping("/admin/themes")
+    public ResponseEntity<CreateThemeResponse> createTheme(
+        @Valid @RequestBody CreateThemeRequest createThemeRequest,
+        HttpServletRequest httpServletRequest
+    ) {
+        if (validator.isUnauthorized(httpServletRequest)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        CreateThemeResponse response = themeService.createTheme(createThemeRequest);
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    }
+
+    @DeleteMapping("/admin/themes/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable Long id, HttpServletRequest request) {
+        if (validator.isUnauthorized(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        themeService.deleteTheme(id);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    }
+}
