+++ src/test/java/roomescape/domain/theme/ThemeServiceTest.java
@@ -6,9 +6,9 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.theme.dto.AdminThemeResponse;
-import roomescape.domain.theme.dto.CreateThemeRequest;
-import roomescape.domain.theme.dto.CreateThemeResponse;
+import roomescape.domain.theme.admin.dto.AdminThemeResponse;
+import roomescape.domain.theme.admin.dto.CreateThemeRequest;
+import roomescape.domain.theme.admin.dto.CreateThemeResponse;
 import roomescape.domain.theme.dto.ThemeResponse;
 import roomescape.support.fake.FakeReservationRepository;
 import roomescape.support.fake.FakeThemeRepository;
